Rename backend-cli/ to agent-cli/ and backend-pi-agent to agent-pi/
