+++ src/agent-cli/index.ts
